+++ CHANGELOG.md
@@ -9,6 +9,11 @@ Wherever possible, a reference to an issue in the project issue tracker should b
 
 ## [Unreleased]
 
+### Fixed
+
+* Corrected layer styles in GeoPackage backups
+  [#160](https://gitlab.data.bas.ac.uk/MAGIC/ops-data-store/-/issues/160)
+
 ## [0.8.0] - 2023-12-13
 
 ### Added